refactor: localStorage 관리 방식 개선 및 타입 에러 수정
- useLocalStorage 훅 제거하고 원본 패턴으로 복귀
- App.tsx에서 중앙 집중식 localStorage 관리
- addNotification 타입 순서 통일
- React Hook 의존성 배열 수정
- 대량 구매 할인 로직 수정
- useProducts 훅 추가로 상품 관리 분리
