--- src/basic/hooks/useLocalStorage.ts
@@ -1,4 +1,0 @@
-// TODO: useLocalStorage
-export function useLocalStorage() {
-  // TODO: 구현
-}
