+++ src/basic/hooks/useProducts.ts
@@ -1,18 +1,69 @@
-// TODO: 상품 관리 Hook
-// 힌트:
-// 1. 상품 목록 상태 관리 (localStorage 연동 고려)
-// 2. 상품 CRUD 작업
-// 3. 재고 업데이트
-// 4. 할인 규칙 추가/삭제
-//
-// 반환할 값:
-// - products: 상품 배열
-// - updateProduct: 상품 정보 수정
-// - addProduct: 새 상품 추가
-// - updateProductStock: 재고 수정
-// - addProductDiscount: 할인 규칙 추가
-// - removeProductDiscount: 할인 규칙 삭제
+import { useCallback, useState } from 'react';
+
+import { ProductWithUI } from '../../types';
+import { initialProducts } from '../constants';
 
 export function useProducts() {
-  // TODO: 구현
+  // localStorage에서 초기값 가져오기 (원본 패턴과 동일)
+  const [products, setProducts] = useState<ProductWithUI[]>(() => {
+    const saved = localStorage.getItem('products');
+    if (saved) {
+      try {
+        return JSON.parse(saved);
+      } catch {
+        return initialProducts;
+      }
+    }
+    return initialProducts;
+  });
+
+  // addProduct 함수
+  const addProduct = useCallback(
+    (
+      newProduct: Omit<ProductWithUI, 'id'>,
+      onNotification?: (message: string, type: 'success' | 'error' | 'warning') => void
+    ) => {
+      const product: ProductWithUI = {
+        ...newProduct,
+        id: `p${Date.now()}`,
+      };
+      setProducts((prev) => [...prev, product]);
+      onNotification?.('상품이 추가되었습니다.', 'success');
+    },
+    []
+  );
+
+  // updateProduct 함수
+  const updateProduct = useCallback(
+    (
+      productId: string,
+      updates: Partial<ProductWithUI>,
+      onNotification?: (message: string, type: 'success' | 'error' | 'warning') => void
+    ) => {
+      setProducts((prev) =>
+        prev.map((product) => (product.id === productId ? { ...product, ...updates } : product))
+      );
+      onNotification?.('상품이 수정되었습니다.', 'success');
+    },
+    []
+  );
+
+  // deleteProduct 함수
+  const deleteProduct = useCallback(
+    (
+      productId: string,
+      onNotification?: (message: string, type: 'success' | 'error' | 'warning') => void
+    ) => {
+      setProducts((prev) => prev.filter((p) => p.id !== productId));
+      onNotification?.('상품이 삭제되었습니다.', 'success');
+    },
+    []
+  );
+
+  return {
+    products,
+    addProduct,
+    updateProduct,
+    deleteProduct,
+  };
 }
