+++ src/basic/models/cart.ts
@@ -25,9 +25,8 @@ export const getMaxApplicableDiscount = (item: CartItem, cart: CartItem[]): numb
       : maxDiscount;
   }, 0);
 
-  // 대량 구매 시 추가 5% 할인 (모든 필요한 데이터를 파라미터로 전달받음)
-  const hasBulkPurchase = cart.some((cartItem) => cartItem.quantity >= 10);
-  if (hasBulkPurchase) {
+  // 대량 구매 시 추가 5% 할인 (해당 상품이 10개 이상일 때)
+  if (quantity >= 10) {
     return Math.min(baseDiscount + 0.05, 0.5);
   }
